PIM-5853: Refactor datetime filter / remove UpdatedDateTimeFilter
