+++ src/Pim/Bundle/CatalogBundle/Resources/config/query_builders.yml
@@ -123,18 +123,20 @@ services:
     pim_catalog.doctrine.query.filter.datetime:
         class: '%pim_catalog.doctrine.query.filter.datetime.class%'
         arguments:
-            - ['created', 'updated']
-            - ['=', '<', '>', 'BETWEEN', 'NOT BETWEEN', 'EMPTY', 'NOT EMPTY', '!=']
-        tags:
-            - { name: 'pim_catalog.doctrine.query.filter', priority: 30 }
-
-    pim_catalog.doctrine.query.filter.updated_datetime:
-        class: '%pim_catalog.doctrine.query.filter.updated_datetime.class%'
-        arguments:
             - '@pim_import_export.repository.job_instance'
             - '@akeneo_batch.job_repository'
-            - ['updated']
-            - ['SINCE LAST N DAYS', 'SINCE LAST EXPORT']
+            - ['created', 'updated']
+            -
+                - '='
+                - '<'
+                - '>'
+                - 'BETWEEN'
+                - 'NOT BETWEEN'
+                - 'EMPTY'
+                - 'NOT EMPTY'
+                - '!='
+                - 'SINCE LAST N DAYS'
+                - 'SINCE LAST EXPORT'
         tags:
             - { name: 'pim_catalog.doctrine.query.filter', priority: 30 }
