+++ src/Pim/Bundle/CatalogBundle/Resources/config/storage_driver/doctrine/orm.yml
@@ -9,21 +9,20 @@ parameters:
     pim_catalog.event_subscriber.orm.inject_product_reference.class: Pim\Bundle\CatalogBundle\EventSubscriber\ORM\InjectProductReferenceSubscriber
     pim_catalog.event_subscriber.metric_base_values.class:           Pim\Bundle\CatalogBundle\EventSubscriber\ORM\MetricBaseValuesSubscriber
 
-    pim_catalog.doctrine.query.filter.product_id.class:       Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\ProductIdFilter
-    pim_catalog.doctrine.query.filter.number.class:           Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\NumberFilter
-    pim_catalog.doctrine.query.filter.string.class:           Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\StringFilter
-    pim_catalog.doctrine.query.filter.boolean.class:          Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\BooleanFilter
-    pim_catalog.doctrine.query.filter.family.class:           Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\FamilyFilter
-    pim_catalog.doctrine.query.filter.groups.class:           Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\GroupsFilter
-    pim_catalog.doctrine.query.filter.date.class:             Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\DateFilter
-    pim_catalog.doctrine.query.filter.datetime.class:         Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\DateTimeFilter
-    pim_catalog.doctrine.query.filter.updated_datetime.class: Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\UpdatedDateTimeFilter
-    pim_catalog.doctrine.query.filter.completeness.class:     Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\CompletenessFilter
-    pim_catalog.doctrine.query.filter.price.class:            Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\PriceFilter
-    pim_catalog.doctrine.query.filter.metric.class:           Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\MetricFilter
-    pim_catalog.doctrine.query.filter.option.class:           Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\OptionFilter
-    pim_catalog.doctrine.query.filter.options.class:          Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\OptionsFilter
-    pim_catalog.doctrine.query.filter.media.class:            Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\MediaFilter
+    pim_catalog.doctrine.query.filter.product_id.class:   Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\ProductIdFilter
+    pim_catalog.doctrine.query.filter.number.class:       Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\NumberFilter
+    pim_catalog.doctrine.query.filter.string.class:       Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\StringFilter
+    pim_catalog.doctrine.query.filter.boolean.class:      Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\BooleanFilter
+    pim_catalog.doctrine.query.filter.family.class:       Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\FamilyFilter
+    pim_catalog.doctrine.query.filter.groups.class:       Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\GroupsFilter
+    pim_catalog.doctrine.query.filter.date.class:         Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\DateFilter
+    pim_catalog.doctrine.query.filter.datetime.class:     Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\DateTimeFilter
+    pim_catalog.doctrine.query.filter.completeness.class: Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\CompletenessFilter
+    pim_catalog.doctrine.query.filter.price.class:        Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\PriceFilter
+    pim_catalog.doctrine.query.filter.metric.class:       Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\MetricFilter
+    pim_catalog.doctrine.query.filter.option.class:       Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\OptionFilter
+    pim_catalog.doctrine.query.filter.options.class:      Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\OptionsFilter
+    pim_catalog.doctrine.query.filter.media.class:        Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\MediaFilter
 
     pim_catalog.doctrine.query.cursor.class:               Akeneo\Bundle\StorageUtilsBundle\Doctrine\ORM\Cursor\Cursor
     pim_catalog.doctrine.cursor_factory.product.class:     Akeneo\Bundle\StorageUtilsBundle\Doctrine\ORM\Cursor\CursorFactory
