+++ src/Pim/Bundle/CatalogBundle/spec/Doctrine/ORM/Filter/DateTimeFilterSpec.php
@@ -2,19 +2,40 @@
 
 namespace spec\Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter;
 
+use Akeneo\Component\Batch\Job\JobRepositoryInterface;
+use Akeneo\Component\Batch\Model\JobExecution;
+use Akeneo\Component\Batch\Model\JobInstance;
 use Doctrine\ORM\Query\Expr;
+use Doctrine\ORM\Query\Expr\Comparison;
 use Doctrine\ORM\QueryBuilder;
 use PhpSpec\ObjectBehavior;
+use Pim\Bundle\ImportExportBundle\Entity\Repository\JobInstanceRepository;
 use Pim\Component\Catalog\Exception\InvalidArgumentException;
 use Prophecy\Argument;
 
 class DateTimeFilterSpec extends ObjectBehavior
 {
-    function let(QueryBuilder $qb)
-    {
+    function let(
+        QueryBuilder $qb,
+        JobInstanceRepository $jobInstanceRepository,
+        JobRepositoryInterface $jobRepository
+    ) {
         $this->beConstructedWith(
+            $jobInstanceRepository,
+            $jobRepository,
             ['created', 'updated'],
-            ['=', '<', '>', 'BETWEEN', 'NOT BETWEEN', 'EMPTY', 'NOT EMPTY', '!=']
+            [
+                '=',
+                '<',
+                '>',
+                'BETWEEN',
+                'NOT BETWEEN',
+                'EMPTY',
+                'NOT EMPTY',
+                '!=',
+                'SINCE LAST EXPORT',
+                'SINCE LAST N DAYS'
+            ]
         );
         $this->setQueryBuilder($qb);
 
@@ -38,7 +59,18 @@ class DateTimeFilterSpec extends ObjectBehavior
 
     function it_supports_operators()
     {
-        $this->getOperators()->shouldReturn(['=', '<', '>', 'BETWEEN', 'NOT BETWEEN', 'EMPTY', 'NOT EMPTY', '!=']);
+        $this->getOperators()->shouldReturn([
+            '=',
+            '<',
+            '>',
+            'BETWEEN',
+            'NOT BETWEEN',
+            'EMPTY',
+            'NOT EMPTY',
+            '!=',
+            'SINCE LAST EXPORT',
+            'SINCE LAST N DAYS'
+        ]);
 
         $this->supportsOperator('=')->shouldReturn(true);
         $this->supportsOperator('FAKE')->shouldReturn(false);
@@ -165,6 +197,111 @@ class DateTimeFilterSpec extends ObjectBehavior
         $qb->andWhere($or)->shouldBeCalled();
 
         $this->addFieldFilter('updated_at', 'NOT BETWEEN', ['2014-03-15 12:03:00', '2014-03-16 12:03:00']);
+    }
+
+    function it_adds_an_updated_since_last_n_days_filter($qb, Expr $expr, Comparison $comparison)
+    {
+        $qb->getRootAliases()->willReturn(['alias']);
+        $qb->andWhere($comparison)->shouldBeCalled();
+        $qb->expr()->willReturn($expr);
+        $expr->gt('alias.updated', Argument::type('string'))->shouldBeCalled()->willReturn($comparison);
+
+        $expr->literal(Argument::type('string'))->shouldBeCalled()->willReturn('2016-06-20 16:42:42');
+        $expr->gt('alias.updated', '2016-06-20 16:42:42')->shouldBeCalled()->willReturn($comparison);
+
+        $this->addFieldFilter(
+            'updated',
+            'SINCE LAST N DAYS',
+            30,
+            null,
+            null
+        );
+    }
+
+    function it_adds_an_updated_since_last_export_filter(
+        $qb,
+        $jobInstanceRepository,
+        $jobRepository,
+        JobInstance $jobInstance,
+        JobExecution $jobExecution,
+        \DateTime $startTime,
+        Expr $expr,
+        Comparison $comparison
+    ) {
+        $jobInstanceRepository->findOneBy(['code' => 'csv_product_export'])->willReturn($jobInstance);
+        $jobRepository->getLastJobExecution($jobInstance, 1)->shouldBeCalled()->willReturn($jobExecution);
+
+        $jobExecution->getStartTime()->willReturn($startTime);
+        $startTime->format('Y-m-d H:i:s')->willReturn('2016-06-20 16:42:42');
+
+        $qb->getRootAliases()->willReturn(['alias']);
+        $qb->expr()->willReturn($expr);
+        $qb->andWhere($comparison)->shouldBeCalled();
+
+        $expr->literal('2016-06-20 16:42:42')->shouldBeCalled()->willReturn('2016-06-20 16:42:42');
+        $expr->gt('alias.updated', '2016-06-20 16:42:42')->shouldBeCalled()->willReturn($comparison);
+
+        $this->addFieldFilter(
+            'updated',
+            'SINCE LAST EXPORT',
+            'csv_product_export',
+            null,
+            null
+        );
+    }
+
+    function it_does_not_add_an_updated_since_last_export_filter_if_no_option_given(
+        $qb,
+        $jobInstanceRepository,
+        $jobRepository,
+        JobInstance $jobInstance
+    ) {
+        $jobInstanceRepository->findOneBy(['code' => 'csv_product_export'])->willReturn($jobInstance);
+        $jobRepository->getLastJobExecution($jobInstance, 1)->shouldBeCalled()->willReturn(null);
+
+        $qb->andWhere(Argument::cetera())->shouldNotBeCalled();
+
+        $this->addFieldFilter(
+            'updated',
+            'SINCE LAST EXPORT',
+            'csv_product_export',
+            null,
+            null
+        );
+    }
+
+    function it_throws_an_exception_if_value_is_not_a_string_for_since_last_export()
+    {
+        $this
+            ->shouldThrow(
+                InvalidArgumentException::stringExpected('updated', 'filter', 'updated', 'integer')
+            )->during(
+                'addFieldFilter',
+                [
+                    'updated',
+                    'SINCE LAST EXPORT',
+                    42,
+                    null,
+                    null,
+                ]
+            );
+    }
+
+    function it_throws_an_exception_if_value_is_not_a_numeric_for_since_last_n_days()
+    {
+        $this
+            ->shouldThrow(
+                InvalidArgumentException::numericExpected('updated', 'filter', 'updated', 'string')
+            )->during(
+                'addFieldFilter',
+                [
+                    'updated',
+                    'SINCE LAST N DAYS',
+                    'csv_product_export',
+                    null,
+                    null
+                ]
+            );
     }
 
     function it_throws_an_exception_if_value_is_not_a_string_an_array_or_a_datetime()
